# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/bookings": {
      "post": {
        "operationId": "BookingController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Bookings"
        ]
      }
    },
    "/bookings/{id}/pay": {
      "post": {
        "operationId": "BookingController_pay",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayBookingBody"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Bookings"
        ]
      }
    },
    "/bookings/{id}/expire": {
      "post": {
        "operationId": "BookingController_expire",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Bookings"
        ]
      }
    },
    "/tickets": {
      "get": {
        "operationId": "TicketController_list",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Tickets"
        ]
      }
    },
    "/tickets/check-in": {
      "post": {
        "operationId": "TicketController_checkIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInTicketBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Tickets"
        ]
      }
    },
    "/events": {
      "post": {
        "operationId": "EventController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      },
      "get": {
        "operationId": "EventController_list",
        "parameters": [
          {
            "name": "date",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "location",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}/publish": {
      "post": {
        "operationId": "EventController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}/cancel": {
      "post": {
        "operationId": "EventController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelEventBody"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}": {
      "get": {
        "operationId": "EventController_details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}/participants": {
      "get": {
        "operationId": "EventController_participants",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}/sales-report": {
      "get": {
        "operationId": "EventController_salesReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{eventId}/ticket-categories": {
      "post": {
        "operationId": "TicketCategoryController_create",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketCategoryBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Ticket Categories"
        ]
      }
    },
    "/events/{eventId}/ticket-categories/{categoryId}/disable": {
      "post": {
        "operationId": "TicketCategoryController_disable",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Ticket Categories"
        ]
      }
    },
    "/refunds": {
      "post": {
        "operationId": "RefundController_request",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestRefundBody"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Refunds"
        ]
      }
    },
    "/refunds/{id}/reject": {
      "post": {
        "operationId": "RefundController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRefundBody"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Refunds"
        ]
      }
    },
    "/refunds/{id}/approve": {
      "post": {
        "operationId": "RefundController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Refunds"
        ]
      }
    },
    "/refunds/{id}/pay-out": {
      "post": {
        "operationId": "RefundController_payOut",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Refunds"
        ]
      }
    },
    "/users": {
      "post": {
        "operationId": "UserController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}/role": {
      "patch": {
        "operationId": "UserController_changeRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeUserRoleBody"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/by-email": {
      "get": {
        "operationId": "UserController_byEmail",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UserController_profile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    }
  },
  "info": {
    "title": "Eventix API",
    "description": "REST API for the Eventix Event Ticketing & Booking System. Endpoints are grouped by bounded context: Users, Events, Ticket Categories, Bookings, Tickets, and Refunds.",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "Users",
      "description": ""
    },
    {
      "name": "Events",
      "description": ""
    },
    {
      "name": "Ticket Categories",
      "description": ""
    },
    {
      "name": "Bookings",
      "description": ""
    },
    {
      "name": "Tickets",
      "description": ""
    },
    {
      "name": "Refunds",
      "description": ""
    }
  ],
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "schemas": {
      "CreateBookingBody": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string"
          },
          "eventId": {
            "type": "string"
          },
          "ticketCategoryId": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          }
        }
      },
      "PayBookingBody": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          }
        }
      },
      "CheckInTicketBody": {
        "type": "object",
        "properties": {
          "ticketCode": {
            "type": "string"
          },
          "eventId": {
            "type": "string"
          }
        }
      },
      "CreateEventBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "maxCapacity": {
            "type": "number"
          }
        }
      },
      "CancelEventBody": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
      "CreateTicketCategoryBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "quota": {
            "type": "number"
          },
          "salesStartDate": {
            "type": "string"
          },
          "salesEndDate": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          }
        }
      },
      "RequestRefundBody": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          }
        }
      },
      "RejectRefundBody": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
      "RegisterUserBody": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
      "ChangeUserRoleBody": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string"
          }
        }
      }
    }
  }
}